Refactor settings page with tabbed design and improved UX
