+++ SETTINGS_REDESIGN.md
@@ -1,0 +1,105 @@
+# Переробка налаштувань - Новий дизайн з вкладками
+
+## Опис змін
+
+Налаштування були перероблені з використанням сучасного дизайну, подібного до GitHub та інших популярних сервісів. Тепер налаштування мають:
+
+- **Вкладки зліва** - навігація між різними розділами налаштувань
+- **Контент справа** - відображення вмісту обраної вкладки
+- **Адаптивний дизайн** - підтримка мобільних пристроїв
+
+## Структура файлів
+
+### Нові компоненти:
+- `frontend/src/pages/settings/SettingsPage.tsx` - основний компонент з вкладками
+- `frontend/src/pages/settings/AccountSettings.tsx` - налаштування акаунту
+- `frontend/src/pages/settings/SessionsSettings.tsx` - управління сесіями
+- `frontend/src/styles/SettingsPage.css` - стилі для нового дизайну
+
+### Видалені файли:
+- `frontend/src/pages/settings/AccountSettingsPage.tsx` (замінено на AccountSettings.tsx)
+- `frontend/src/pages/settings/SessionsPage.tsx` (замінено на SessionsSettings.tsx)
+
+## Доступні вкладки
+
+### 1. Account (👤)
+- Інформація про користувача
+- Налаштування браузерного розширення
+- Налаштування домену для карток
+
+### 2. Sessions (🔐)
+- Управління активними сесіями
+- Перегляд та відкликання сесій з інших пристроїв
+- Статус поточної сесії
+
+## Маршрутизація
+
+Нова структура маршрутів:
+- `/settings` - головна сторінка налаштувань (вкладка Account)
+- `/settings/sessions` - управління сесіями
+
+## Особливості дизайну
+
+### Desktop версія:
+- Бічна панель з вкладками зліва (240px ширина)
+- Основний контент справа
+- Максимальна ширина контейнера: 1200px
+
+### Mobile версія:
+- Вкладки відображаються горизонтально згортаються
+- Контент займає повну ширину екрану
+- Адаптивні елементи інтерфейсу
+
+## Стилізація
+
+### Кольорова схема:
+- Основний текст: `#24292e`
+- Другорядний текст: `#586069`
+- Активна вкладка: `#0366d6` з фоном `#f1f8ff`
+- Рамки: `#e1e4e8`
+- Фон: `#ffffff`
+
+### Інтерактивні елементи:
+- Hover ефекти для вкладок
+- Focus стани для полів вводу
+- Анімації переходів (0.2s ease)
+
+## Використання
+
+### Навігація:
+```typescript
+// Перехід між вкладками
+navigate('/settings'); // Account вкладка
+navigate('/settings/sessions'); // Sessions вкладка
+```
+
+### Додавання нової вкладки:
+1. Створіть новий компонент для контенту вкладки
+2. Додайте вкладку в масив `tabs` в `SettingsPage.tsx`
+3. Додайте маршрут в `Routes`
+4. Створіть відповідні стилі
+
+## Переваги нового дизайну
+
+1. **Кращий UX** - звичний інтерфейс для користувачів
+2. **Масштабованість** - легко додавати нові розділи налаштувань
+3. **Консистентність** - єдиний стиль для всіх налаштувань
+4. **Адаптивність** - працює на всіх пристроях
+5. **Продуктивність** - швидка навігація між розділами
+
+## Технічні деталі
+
+### React Router:
+- Використовується `useNavigate` для навігації
+- `useLocation` для відстеження поточного шляху
+- `useEffect` для синхронізації стану з URL
+
+### State Management:
+- Локальний стан для активної вкладки
+- Синхронізація з URL через `useEffect`
+
+### CSS:
+- Flexbox для layout
+- CSS Grid для складних компонентів
+- CSS Custom Properties для кольорів
+- Media queries для адаптивності
