+++ frontend/src/pages/settings/index.ts
@@ -1,2 +1,1 @@
-export { default as AccountSettingsPage } from './AccountSettingsPage';
-export { default as SessionsPage } from './SessionsPage'; 
+export { default as SettingsPage } from './SettingsPage';
