+++ frontend/src/styles/SettingsPage.css
@@ -1,0 +1,396 @@
+/* Settings Page Layout */
+.settings-page {
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 20px;
+}
+
+.settings-header {
+  margin-bottom: 30px;
+  border-bottom: 1px solid #e1e4e8;
+  padding-bottom: 20px;
+}
+
+.settings-header h1 {
+  margin: 0;
+  font-size: 24px;
+  font-weight: 600;
+  color: #24292e;
+}
+
+.settings-layout {
+  display: flex;
+  gap: 40px;
+  min-height: 600px;
+}
+
+/* Sidebar Navigation */
+.settings-sidebar {
+  width: 240px;
+  flex-shrink: 0;
+}
+
+.settings-nav {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.settings-nav-item {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  padding: 12px 16px;
+  text-decoration: none;
+  color: #586069;
+  border-radius: 6px;
+  font-size: 14px;
+  font-weight: 500;
+  transition: all 0.2s ease;
+  border: 1px solid transparent;
+}
+
+.settings-nav-item:hover {
+  background-color: #f6f8fa;
+  color: #24292e;
+  text-decoration: none;
+}
+
+.settings-nav-item.active {
+  background-color: #f1f8ff;
+  color: #0366d6;
+  border-color: #c8e1ff;
+}
+
+.nav-icon {
+  font-size: 16px;
+  width: 20px;
+  text-align: center;
+}
+
+.nav-label {
+  font-weight: 500;
+}
+
+/* Content Area */
+.settings-content {
+  flex: 1;
+  min-width: 0;
+}
+
+/* Settings Sections */
+.settings-section {
+  margin-bottom: 40px;
+}
+
+.settings-section h2 {
+  margin: 0 0 16px 0;
+  font-size: 20px;
+  font-weight: 600;
+  color: #24292e;
+  border-bottom: 1px solid #e1e4e8;
+  padding-bottom: 8px;
+}
+
+.settings-section h3 {
+  margin: 0 0 12px 0;
+  font-size: 16px;
+  font-weight: 600;
+  color: #24292e;
+}
+
+.settings-section h4 {
+  margin: 0 0 8px 0;
+  font-size: 14px;
+  font-weight: 600;
+  color: #24292e;
+}
+
+/* Setting Items */
+.setting-item {
+  margin-bottom: 24px;
+}
+
+.setting-item label {
+  display: block;
+  margin-bottom: 8px;
+  font-weight: 600;
+  color: #24292e;
+  font-size: 14px;
+}
+
+.setting-value {
+  padding: 8px 12px;
+  background-color: #f6f8fa;
+  border: 1px solid #e1e4e8;
+  border-radius: 6px;
+  font-size: 14px;
+  color: #24292e;
+  min-height: 20px;
+}
+
+.setting-description {
+  margin-bottom: 12px;
+  color: #586069;
+  font-size: 14px;
+  line-height: 1.5;
+}
+
+.setting-input {
+  width: 100%;
+  max-width: 400px;
+  padding: 8px 12px;
+  border: 1px solid #e1e4e8;
+  border-radius: 6px;
+  font-size: 14px;
+  background-color: #ffffff;
+  transition: border-color 0.2s ease;
+}
+
+.setting-input:focus {
+  outline: none;
+  border-color: #0366d6;
+  box-shadow: 0 0 0 3px rgba(3, 102, 214, 0.1);
+}
+
+/* Buttons */
+.button {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  padding: 8px 16px;
+  border: 1px solid transparent;
+  border-radius: 6px;
+  font-size: 14px;
+  font-weight: 500;
+  text-decoration: none;
+  cursor: pointer;
+  transition: all 0.2s ease;
+  background: none;
+  font-family: inherit;
+}
+
+.button-secondary {
+  color: #24292e;
+  background-color: #ffffff;
+  border-color: #e1e4e8;
+}
+
+.button-secondary:hover {
+  background-color: #f6f8fa;
+  border-color: #d1d5da;
+}
+
+.button-danger {
+  color: #ffffff;
+  background-color: #d73a49;
+  border-color: #d73a49;
+}
+
+.button-danger:hover {
+  background-color: #cb2431;
+  border-color: #cb2431;
+}
+
+/* Messages */
+.success-message {
+  padding: 12px 16px;
+  background-color: #d4edda;
+  border: 1px solid #c3e6cb;
+  border-radius: 6px;
+  color: #155724;
+  font-size: 14px;
+  margin-bottom: 20px;
+}
+
+.error-message {
+  padding: 12px 16px;
+  background-color: #f8d7da;
+  border: 1px solid #f5c6cb;
+  border-radius: 6px;
+  color: #721c24;
+  font-size: 14px;
+  margin-bottom: 20px;
+  display: flex;
+  align-items: center;
+  gap: 12px;
+}
+
+/* Sessions Specific Styles */
+.sessions-content {
+  width: 100%;
+}
+
+.sessions-list {
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  margin-bottom: 24px;
+}
+
+.session-item {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 16px;
+  border: 1px solid #e1e4e8;
+  border-radius: 6px;
+  background-color: #ffffff;
+}
+
+.session-item.current-session {
+  background-color: #f1f8ff;
+  border-color: #c8e1ff;
+}
+
+.session-header {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  margin-bottom: 8px;
+}
+
+.session-id {
+  font-weight: 600;
+  color: #24292e;
+  font-size: 14px;
+}
+
+.current-badge {
+  background-color: #28a745;
+  color: #ffffff;
+  padding: 2px 8px;
+  border-radius: 12px;
+  font-size: 12px;
+  font-weight: 500;
+}
+
+.session-details {
+  color: #586069;
+  font-size: 13px;
+}
+
+.session-details p {
+  margin: 2px 0;
+}
+
+/* Feature Coming Soon */
+.feature-coming-soon {
+  padding: 24px;
+  border: 1px solid #e1e4e8;
+  border-radius: 6px;
+  background-color: #f6f8fa;
+  text-align: center;
+  margin-bottom: 24px;
+}
+
+.feature-icon {
+  font-size: 48px;
+  margin-bottom: 16px;
+}
+
+.feature-description {
+  color: #586069;
+  font-size: 14px;
+  margin-bottom: 20px;
+}
+
+.feature-benefits {
+  text-align: left;
+  margin-bottom: 20px;
+}
+
+.feature-benefits ul {
+  margin: 8px 0;
+  padding-left: 20px;
+}
+
+.feature-benefits li {
+  color: #586069;
+  font-size: 14px;
+  margin-bottom: 4px;
+}
+
+.feature-status {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  align-items: center;
+}
+
+.status-item {
+  display: flex;
+  gap: 8px;
+  font-size: 14px;
+}
+
+.status-label {
+  color: #586069;
+}
+
+.status-value {
+  font-weight: 500;
+}
+
+.status-value.ready {
+  color: #28a745;
+}
+
+.status-value.pending {
+  color: #f6a434;
+}
+
+/* Actions */
+.settings-actions {
+  display: flex;
+  gap: 12px;
+  margin-top: 24px;
+}
+
+/* Loading and Empty States */
+.loading {
+  text-align: center;
+  padding: 40px;
+  color: #586069;
+  font-size: 14px;
+}
+
+.no-sessions {
+  text-align: center;
+  padding: 40px;
+  color: #586069;
+  font-size: 14px;
+}
+
+/* Responsive Design */
+@media (max-width: 768px) {
+  .settings-layout {
+    flex-direction: column;
+    gap: 20px;
+  }
+  
+  .settings-sidebar {
+    width: 100%;
+  }
+  
+  .settings-nav {
+    flex-direction: row;
+    overflow-x: auto;
+    padding-bottom: 8px;
+  }
+  
+  .settings-nav-item {
+    white-space: nowrap;
+    flex-shrink: 0;
+  }
+  
+  .session-item {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 12px;
+  }
+  
+  .settings-actions {
+    flex-direction: column;
+  }
+}
